Moved Gradle Hints and CLI editor into a new module
