+++ extide/gradle.editor/manifest.mf
@@ -1,0 +1,5 @@
+Manifest-Version: 1.0
+AutoUpdate-Show-In-Client: true
+OpenIDE-Module: org.netbeans.modules.gradle.editor
+OpenIDE-Module-Localizing-Bundle: org/netbeans/modules/gradle/editor/Bundle.properties
+OpenIDE-Module-Specification-Version: 1.0
