+++ extide/gradle.editor/src/org/netbeans/modules/gradle/editor/MimeTypes.java
@@ -1,0 +1,33 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.netbeans.modules.gradle.editor;
+
+/**
+ *
+ * @author lkishalmi
+ */
+public final class MimeTypes {
+
+    public static final String GRADLE_CLI = "text/x-gradle-cli"; //NOI18N
+    public static final String GRADLE_FILE = "text/x-gradle+x-groovy"; //NOI18N
+    public static final String GRADLE_KOTLIN_FILE = "text/x-gradle+x-kotlin"; //NOI18N
+
+    private MimeTypes() {}
+}
+
