+++ extide/gradle.editor/src/org/netbeans/modules/gradle/editor/cli/GradleCliEditorKit.java
@@ -16,24 +16,23 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.netbeans.modules.gradle.execute;
 
-import org.junit.Test;
-import static org.junit.Assert.*;
+package org.netbeans.modules.gradle.editor.cli;
+
+import org.netbeans.modules.editor.NbEditorKit;
 
 /**
  *
  * @author lkishalmi
  */
-public class GradleCliCompletionProviderTest {
 
-    /**
-     * Test of createTask method, of class GradleCliCompletionProvider.
-     */
-    @Test
-    public void testCreateInstance() {
-        GradleCliCompletionProvider instance = new GradleCliCompletionProvider();
-        assertNotNull(instance);
+public class GradleCliEditorKit extends NbEditorKit {
+
+    public static final String MIME_TYPE = "text/x-gradle-cli"; //NOI18N
+
+    @Override
+    public String getContentType() {
+        return MIME_TYPE; 
     }
 
 }
