+++ extide/gradle.editor/src/org/netbeans/modules/gradle/editor/hints/GradleHintsProvider.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.netbeans.modules.gradle.editor;
+package org.netbeans.modules.gradle.editor.hints;
 
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
