+++ extide/gradle/src/org/netbeans/modules/gradle/configurations/NewConfigurationPanel.java
@@ -31,7 +31,6 @@ import javax.swing.text.Document;
 import javax.swing.text.EditorKit;
 import org.netbeans.api.project.Project;
 import org.netbeans.modules.gradle.api.execute.GradleExecConfiguration;
-import org.netbeans.modules.gradle.execute.GradleCliEditorKit;
 import org.openide.NotificationLineSupport;
 import org.openide.filesystems.FileUtil;
 import org.openide.text.CloneableEditorSupport;
@@ -54,7 +53,7 @@ public class NewConfigurationPanel extends javax.swing.JPanel implements Documen
         this.configProvider = confProvider;
 
         initComponents();
-        EditorKit kit = CloneableEditorSupport.getEditorKit(GradleCliEditorKit.MIME_TYPE);
+        EditorKit kit = CloneableEditorSupport.getEditorKit("text/x-gradle-cli"); //NOI18N
         txParameters.setEditorKit(kit);
         txParameters.getDocument().putProperty(Document.StreamDescriptionProperty, project);
         if (isNew) {
