+++ nbbuild/cluster.properties
@@ -1492,6 +1492,7 @@ nb.cluster.extide.depends=\
 nb.cluster.extide=\
         gradle,\
         gradle.dists,\
+        gradle.editor,\
         libs.gradle,\
         o.apache.tools.ant.module,\
         options.java
